+++ src/daemon/gateway.go
@@ -493,8 +493,8 @@ func (gw *Gateway) InjectBroadcastTransaction(txn coin.Transaction) error {
 	var err error
 	gw.strand("InjectBroadcastTransaction", func() {
 		err = gw.v.WithUpdateTx("gateway.InjectBroadcastTransaction", func(tx *dbutil.Tx) error {
-			if _, err := gw.v.InjectTransactionUserTx(tx, txn); err != nil {
-				logger.WithError(err).Error("InjectUserTransaction failed")
+			if _, err := gw.v.InjectUserTransactionTx(tx, txn); err != nil {
+				logger.WithError(err).Error("InjectUserTransactionTx failed")
 				return err
 			}