+++ src/visor/blockchain_verify_test.go
@@ -347,11 +347,11 @@ func TestVerifyTransactionSoftHardConstraints(t *testing.T) {
 	// create normal spending txn
 	uxs := coin.CreateUnspents(gb.Head, gb.Body.Transactions[0])
 	txn := makeSpendTx(t, uxs, []cipher.SecKey{genSecret}, toAddr, coins)
-	err = verifySingleTxnSoftHardConstraints(txn, params.MaxUserTransactionSize, params.CoinHourBurnFactor)
+	err = verifySingleTxnSoftHardConstraints(txn, params.MaxUserTransactionSize, params.UserBurnFactor)
 	require.NoError(t, err)
 
 	// Transaction size exceeds maxSize
-	err = verifySingleTxnSoftHardConstraints(txn, txn.Size()-1, params.CoinHourBurnFactor)
+	err = verifySingleTxnSoftHardConstraints(txn, txn.Size()-1, params.UserBurnFactor)
 	requireSoftViolation(t, "Transaction size bigger than max block size", err)
 
 	// Invalid transaction fee
@@ -361,12 +361,12 @@ func TestVerifyTransactionSoftHardConstraints(t *testing.T) {
 		hours += ux.Body.Hours
 	}
 	txn = makeSpendTxWithHoursBurned(t, uxs, []cipher.SecKey{genSecret}, toAddr, coins, 0)
-	err = verifySingleTxnSoftHardConstraints(txn, params.MaxUserTransactionSize, params.CoinHourBurnFactor)
+	err = verifySingleTxnSoftHardConstraints(txn, params.MaxUserTransactionSize, params.UserBurnFactor)
 	requireSoftViolation(t, "Transaction has zero coinhour fee", err)
 
 	// Invalid transaction fee, part 2
 	txn = makeSpendTxWithHoursBurned(t, uxs, []cipher.SecKey{genSecret}, toAddr, coins, 1)
-	err = verifySingleTxnSoftHardConstraints(txn, params.MaxUserTransactionSize, params.CoinHourBurnFactor)
+	err = verifySingleTxnSoftHardConstraints(txn, params.MaxUserTransactionSize, params.UserBurnFactor)
 	requireSoftViolation(t, "Transaction coinhour fee minimum not met", err)
 
 	// Transaction locking is tested by TestVerifyTransactionIsLocked
@@ -374,7 +374,7 @@ func TestVerifyTransactionSoftHardConstraints(t *testing.T) {
 	// Test invalid header hash
 	originInnerHash := txn.InnerHash
 	txn.InnerHash = cipher.SHA256{}
-	err = verifySingleTxnSoftHardConstraints(txn, params.MaxUserTransactionSize, params.CoinHourBurnFactor)
+	err = verifySingleTxnSoftHardConstraints(txn, params.MaxUserTransactionSize, params.UserBurnFactor)
 	requireHardViolation(t, "InnerHash does not match computed hash", err)
 
 	// Set back the originInnerHash
@@ -401,7 +401,7 @@ func TestVerifyTransactionSoftHardConstraints(t *testing.T) {
 	require.NoError(t, err)
 
 	// A UxOut does not exist, it was already spent
-	err = verifySingleTxnSoftHardConstraints(txn, params.MaxUserTransactionSize, params.CoinHourBurnFactor)
+	err = verifySingleTxnSoftHardConstraints(txn, params.MaxUserTransactionSize, params.UserBurnFactor)
 	expectedErr := NewErrTxnViolatesHardConstraint(blockdb.NewErrUnspentNotExist(txn.In[0].Hex()))
 	require.Equal(t, expectedErr, err)
 
@@ -410,21 +410,21 @@ func TestVerifyTransactionSoftHardConstraints(t *testing.T) {
 	_, key := cipher.GenerateKeyPair()
 	toAddr2 := testutil.MakeAddress()
 	tx2 := makeSpendTx(t, uxs, []cipher.SecKey{key, key}, toAddr2, 5e6)
-	err = verifySingleTxnSoftHardConstraints(tx2, params.MaxUserTransactionSize, params.CoinHourBurnFactor)
+	err = verifySingleTxnSoftHardConstraints(tx2, params.MaxUserTransactionSize, params.UserBurnFactor)
 	requireHardViolation(t, "Signature not valid for output being spent", err)
 
 	// Create lost coin transaction
 	uxs2 := coin.CreateUnspents(b.Head, txn)
 	toAddr3 := testutil.MakeAddress()
 	lostCoinTx := makeLostCoinTx(coin.UxArray{uxs2[1]}, []cipher.SecKey{genSecret}, toAddr3, 10e5)
-	err = verifySingleTxnSoftHardConstraints(lostCoinTx, params.MaxUserTransactionSize, params.CoinHourBurnFactor)
+	err = verifySingleTxnSoftHardConstraints(lostCoinTx, params.MaxUserTransactionSize, params.UserBurnFactor)
 	requireHardViolation(t, "Transactions may not destroy coins", err)
 
 	// Create transaction with duplicate UxOuts
 	uxs = coin.CreateUnspents(b.Head, txn)
 	toAddr4 := testutil.MakeAddress()
 	dupUxOutTx := makeDuplicateUxOutTx(coin.UxArray{uxs[0]}, []cipher.SecKey{genSecret}, toAddr4, 1e6)
-	err = verifySingleTxnSoftHardConstraints(dupUxOutTx, params.MaxUserTransactionSize, params.CoinHourBurnFactor)
+	err = verifySingleTxnSoftHardConstraints(dupUxOutTx, params.MaxUserTransactionSize, params.UserBurnFactor)
 	requireHardViolation(t, "Duplicate output in transaction", err)
 }
 
@@ -475,7 +475,7 @@ func TestVerifyTxnFeeCoinHoursAdditionFails(t *testing.T) {
 	testutil.RequireError(t, coinHoursErr, "UxOut.CoinHours addition of earned coin hours overflow")
 
 	// VerifySingleTxnSoftConstraints should fail on this, when trying to calculate the TransactionFee
-	err = VerifySingleTxnSoftConstraints(txn, head.Time()+1e6, uxIn, params.MaxUserTransactionSize, params.CoinHourBurnFactor)
+	err = VerifySingleTxnSoftConstraints(txn, head.Time()+1e6, uxIn, params.MaxUserTransactionSize, params.UserBurnFactor)
 	testutil.RequireError(t, err, NewErrTxnViolatesSoftConstraint(coinHoursErr).Error())
 
 	// VerifySingleTxnHardConstraints should fail on this, when performing the extra check of
@@ -553,7 +553,7 @@ func testVerifyTransactionAddressLocking(t *testing.T, toAddr string, expectedEr
 	})
 	require.NoError(t, err)
 
-	err = VerifySingleTxnSoftConstraints(txn, head.Time(), uxIn, params.MaxUserTransactionSize, params.CoinHourBurnFactor)
+	err = VerifySingleTxnSoftConstraints(txn, head.Time(), uxIn, params.MaxUserTransactionSize, params.UserBurnFactor)
 	if expectedErr == nil {
 		require.NoError(t, err)
 	} else {